director/requirements.txt
ollama>=0.1.0
websockets>=12.0
pydantic>=2.0
httpx>=0.25.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
rich>=13.7.0
